Remove duplicate hooks and consolidate imports
