+++ ESTRUTURA_PROJETO.md
@@ -23,6 +23,9 @@ gerador-ordem-do-dia-brick/
 │   │   │   ├── use-call-sheet.tsx
 │   │   │   ├── use-templates.tsx
 │   │   │   ├── use-call-sheet-history.tsx
+│   │   │   ├── use-projects.ts
+│   │   │   ├── use-project-call-sheets.tsx
+│   │   │   ├── use-sync-storage.ts
 │   │   │   ├── use-toast.ts
 │   │   │   └── use-mobile.tsx
 │   │   ├── 📁 lib/              # Utilitários
